Remove PlcMapper system
* Moves PlcMapper / Tag<M,T> system from an official class to an example.
* Add an alternative implementation for code reuse that uses inheritance.
* Clean up Examples to reflect this change.
* Remove a layer of indirection from Tag / NativeTagWrapper (directly expose NativeTagWrapper as Tag)
