+++ src/Examples/CSharp DotNetCore/ExampleAsync.cs
@@ -6,7 +6,6 @@
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 using libplctag;
-using libplctag.DataTypes;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -21,7 +20,7 @@ namespace CSharpDotNetCore
     {
         public static async Task Run()
         {
-            var myTag = new Tag<DintPlcMapper, int>()
+            var myTag = new Tag()
             {
                 Name = "PROGRAM:SomeProgram.SomeDINT",
                 Gateway = "10.10.10.10",
@@ -33,13 +32,13 @@ namespace CSharpDotNetCore
 
             await myTag.InitializeAsync();
 
-            myTag.Value = 3737;
+            myTag.SetInt32(0, 3737);
 
             await myTag.WriteAsync();
 
             await myTag.ReadAsync();
 
-            int myDint = myTag.Value;
+            int myDint = myTag.GetInt32(0);
 
             Console.WriteLine(myDint);
         }
@@ -50,13 +49,13 @@ namespace CSharpDotNetCore
 
             Console.WriteLine("This method measures the speed of synchronous vs asynchronous reads");
 
-            List<Tag<DintPlcMapper, int>> myTags;
+            List<Tag> myTags;
 
             for (int ii = 0; ii < 10; ii++)
             {
                 myTags = Enumerable.Range(0, 10)
                 .Select(i => {
-                    var myTag = new Tag<DintPlcMapper, int>()
+                    var myTag = new Tag()
                     {
                         Name = $"MY_DINT_ARRAY_1000[{i}]",
                         Gateway = "10.10.10.10",
@@ -111,7 +110,7 @@ namespace CSharpDotNetCore
         {
 
             Console.WriteLine("This method measures the speed of synchronous vs asynchronous reads");
-            var myTag = new Tag<DintPlcMapper, int>()
+            var myTag = new Tag()
             {
                 Name = "PROGRAM:SomeProgram.SomeDINT",
                 Gateway = "10.10.10.10",
@@ -191,7 +190,7 @@ namespace CSharpDotNetCore
 
             var myTags = Enumerable.Range(0, maxTags)
                 .Select(i => {
-                    var myTag = new Tag<DintPlcMapper, int>()
+                    var myTag = new Tag()
                     {
                         Name = "PROGRAM:SomeProgram.SomeDINT",
                         Gateway = "10.10.10.10",
@@ -234,7 +233,7 @@ namespace CSharpDotNetCore
 
             var myTags = Enumerable.Range(0, maxTags)
                 .Select(i => {
-                    var myTag = new Tag<DintPlcMapper, int>()
+                    var myTag = new Tag()
                     {
                         Name = $"MY_DINT_ARRAY_1000[{i}]",
                         Gateway = "10.10.10.10",
