+++ src/Examples/CSharp DotNetCore/Program.cs
@@ -13,20 +13,8 @@ namespace CSharpDotNetCore
     {
         static void Main(string[] args)
         {
-            ILoggerExample.Run();
-            //LoggingExample.Run();
-            //TestDatatypes.Run();
-            //ExampleGenericTag.UDT_Array();
-            //ExampleAsync.SyncAsyncMultipleTagComparison();
-            //ExampleAsync.AsyncParallelCancellation();
-            //ExampleGenericTag.Run();
-            //ExampleRW.Run();
-            //ExampleArray.Run();
-            //ExampleListTags.Run();
-            //ListUdtDefinitions.Run();
-            //ListPlc.Run();
-            //ExampleRW.Run();
-            //ExampleArray.Run();
+            ExampleSimple.Run();
+            ExampleListPlc.Run();
             Console.ReadKey();
         }
     }
