+++ src/Examples/CSharp DotNetFramework/CSharp DotNetFramework.csproj
@@ -43,8 +43,6 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ExampleArray.cs" />
-    <Compile Include="ExampleRW.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
